Initial Application refactor
