+++ Ryujinx.HLE/HOS/Services/Ns/IReadOnlyApplicationControlDataInterface.cs
@@ -13,7 +13,7 @@
 
             long position = context.Request.ReceiveBuff[0].Position;
 
-            byte[] nacpData = context.Device.System.ControlData.ByteSpan.ToArray();
+            byte[] nacpData = context.Device.Application.ControlData.ByteSpan.ToArray();
 
             context.Memory.Write((ulong)position, nacpData);
