+++ Ryujinx.HLE/Switch.cs
@@ -27,6 +27,8 @@ namespace Ryujinx.HLE
 
         public Horizon System { get; private set; }
 
+        public ApplicationLoader Application {get; private set;}
+
         public PerformanceStatistics Statistics { get; private set; }
 
         public Hid Hid { get; private set; }
@@ -59,6 +61,8 @@ namespace Ryujinx.HLE
 
             Hid = new Hid(this, System.HidBaseAddress);
             Hid.InitDevices();
+
+            Application = new ApplicationLoader(this, fileSystem, contentManager);
         }
 
         public void Initialize()
@@ -92,27 +96,27 @@ namespace Ryujinx.HLE
 
         public void LoadCart(string exeFsDir, string romFsFile = null)
         {
-            System.LoadCart(exeFsDir, romFsFile);
+            Application.LoadCart(exeFsDir, romFsFile);
         }
 
         public void LoadXci(string xciFile)
         {
-            System.LoadXci(xciFile);
+            Application.LoadXci(xciFile);
         }
 
         public void LoadNca(string ncaFile)
         {
-            System.LoadNca(ncaFile);
+            Application.LoadNca(ncaFile);
         }
 
         public void LoadNsp(string nspFile)
         {
-            System.LoadNsp(nspFile);
+            Application.LoadNsp(nspFile);
         }
 
         public void LoadProgram(string fileName)
         {
-            System.LoadProgram(fileName);
+            Application.LoadProgram(fileName);
         }
 
         public bool WaitFifo()
